# Improve Token layout and reduce stack frame size for hot paths