--- a/src/Esprima/ParseError.cs
+++ b/src/Esprima/ParseError.cs
@@ -1,6 +1,4 @@
-﻿using static Esprima.EsprimaExceptionHelper;
-
-namespace Esprima;
+﻿namespace Esprima;
 
 public sealed class ParseError
 {