… src/Esprima/JsxToken.cs
… src/Esprima/Scanner.cs
… src/Esprima/Token.cs
… src/Esprima/Utils/JsonTextWriter.cs
… test/Esprima.Tests.SourceGenerators/Snapshots/StringMatcherGeneratorTests.ScannerStringMatchingGeneration.01StringMatchers.g.verified.cs
… test/Esprima.Tests/JavaScriptParserTests.cs
… test/Esprima.Tests/ParserTests.cs
… test/Esprima.Tests/PositionTests.cs