This might the initial cure for #326 , future improvements also possible. Addin some progress here with some steps.

* changed  `FirstCoverInitializedNameError` to object so when frequently referenced, only object pointer required
* slimmed down Token size by using special types for rarely used member encapsulation
* going through some recursive methods and reducing work done in "usual cases" where no complex parsing is needed, splitting methods to smaller units etc

Using a test program how recursive we can get:

```c#
var parser = new JavaScriptParser(new ParserOptions { MaxAssignmentDepth = 1000 });
var depth = 1;
for (depth = 180; depth < 600; ++depth)
{
    var input = $"if ({new string('(', depth)}true{new string(')', depth)}) {{ }}";
    parser.ParseScript(input);
    Console.WriteLine("OK: " + depth);
}
```

Starting point for depth before stack overflow is `253` (Release mode).

fixes #326