@lahma 2022-08-24
Old `Token`:

```
Type layout for 'Token'
Size: 40 bytes. Paddings: 2 bytes (%5 of empty space)
|=============================================|
|   0-7: Object Value (8 bytes)               |
|---------------------------------------------|
|  8-15: Object _customValue (8 bytes)        |
|---------------------------------------------|
| 16-19: Int32 Start (4 bytes)                |
|---------------------------------------------|
| 20-23: Int32 End (4 bytes)                  |
|---------------------------------------------|
| 24-27: Int32 LineNumber (4 bytes)           |
|---------------------------------------------|
| 28-31: Int32 LineStart (4 bytes)            |
|---------------------------------------------|
| 32-33: Char NotEscapeSequenceHead (2 bytes) |
|---------------------------------------------|
|    34: TokenType Type (1 byte)              |
|---------------------------------------------|
|    35: Boolean Octal (1 byte)               |
|---------------------------------------------|
|    36: Boolean Head (1 byte)                |
|---------------------------------------------|
|    37: Boolean Tail (1 byte)                |
|---------------------------------------------|
| 38-39: padding (2 bytes)                    |
|=============================================|
```

New `Token`:
```` 
Type layout for 'Token'
Size: 24 bytes. Paddings: 0 bytes (%0 of empty space)
|===================================|
|   0-7: Object _value (8 bytes)    |
|-----------------------------------|
|  8-11: Int32 Start (4 bytes)      |
|-----------------------------------|
| 12-15: Int32 LineNumber (4 bytes) |
|-----------------------------------|
| 16-19: Int32 LineStart (4 bytes)  |
|-----------------------------------|
| 20-21: UInt16 Length (2 bytes)    |
|-----------------------------------|
|    22: TokenType Type (1 byte)    |
|-----------------------------------|
|    23: Boolean Octal (1 byte)     |
|===================================|

```