@adams85 2022-08-24
Wow, this looks impressive! A max depth of 4-500 should be safe enough. And that 10% improvement in execution time is just the icing on the cake!