@adams85 2022-08-24
BTW, what tool did you use to generate those memory layout diagrams?