@lahma 2022-08-24
> BTW, what tool did you use to generate those memory layout diagrams?

It's https://github.com/SergeyTeplyakov/ObjectLayoutInspector