@lahma 2022-08-25
Now new max depth for test case is `514` in Release mode, created test case to avoid regressions / catch them earlier.

Benchmarks:

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|9.722 ms|0.0205 ms|250.0000|109.3750|4,314 KB|
| **New** |	|	| **8.852 ms (-9%)** | **0.0155 ms** | **250.0000 (0%)** | **109.3750 (0%)** | **4,313 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|1.337 ms|0.0080 ms|41.0156|17.5781|676 KB|
| **New** |	|	| **1.187 ms (-11%)** | **0.0030 ms** | **41.0156 (0%)** | **17.5781 (0%)** | **676 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|7.729 ms|0.0096 ms|226.5625|78.1250|3,765 KB|
| **New** |	|	| **7.030 ms (-9%)** | **0.0102 ms** | **226.5625 (0%)** | **78.1250 (0%)** | **3,761 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|11.767 ms|0.0429 ms|343.7500|171.8750|5,800 KB|
| **New** |	|	| **10.728 ms (-9%)** | **0.0285 ms** | **343.7500 (0%)** | **171.8750 (0%)** | **5,794 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|6.315 ms|0.0117 ms|187.5000|78.1250|3,090 KB|
| **New** |	|	| **5.670 ms (-10%)** | **0.0178 ms** | **187.5000 (0%)** | **78.1250 (0%)** | **3,089 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|1.133 ms|0.0034 ms|33.2031|13.6719|571 KB|
| **New** |	|	| **1.009 ms (-11%)** | **0.0031 ms** | **33.2031 (0%)** | **13.6719 (0%)** | **571 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|5.658 ms|0.0109 ms|171.8750|70.3125|2,857 KB|
| **New** |	|	| **5.087 ms (-10%)** | **0.0107 ms** | **171.8750 (0%)** | **70.3125 (0%)** | **2,854 KB (0%)** |