@sebastienros 2022-08-25
All this perf just by changing the struct?
Is there any potential gains like this for Fluid/Parlot?