@lahma 2022-08-25
It's combination of struct size, usage and the fact that esprima does a lot of deep nested calls. Stack frames were fat and method invocation was more costly.

I went through nesting path and checked methods IL in live view using rider and moved untalented (more rare) paths to own methods so that they would reserve yet more stack space in some branch of the main method. 

Method signature tells all possible needed stack reservation and it adds up with paths never taken.

Don't you try to nerd snipe me towards more parsers!😄