@lahma 2022-08-26
Did the requested changes, due to `Length` change back to `int` the struct got bigger and small regressions, new max for test code is `471` 

## Layout

```
Type layout for 'Token'
Size: 32 bytes. Paddings: 6 bytes (%18 of empty space)
|===================================|
|   0-7: Object _value (8 bytes)    |
|-----------------------------------|
|  8-11: Int32 Start (4 bytes)      |
|-----------------------------------|
| 12-15: Int32 End (4 bytes)        |
|-----------------------------------|
| 16-19: Int32 LineNumber (4 bytes) |
|-----------------------------------|
| 20-23: Int32 LineStart (4 bytes)  |
|-----------------------------------|
|    24: TokenType Type (1 byte)    |
|-----------------------------------|
|    25: Boolean Octal (1 byte)     |
|-----------------------------------|
| 26-31: padding (6 bytes)          |
|===================================|
```

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|9.710 ms|0.0190 ms|250.0000|109.3750|4,314 KB|
| **New** |	|	| **9.040 ms (-7%)** | **0.0201 ms** | **250.0000 (0%)** | **109.3750 (0%)** | **4,315 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|1.360 ms|0.0052 ms|41.0156|17.5781|676 KB|
| **New** |	|	| **1.238 ms (-9%)** | **0.0018 ms** | **41.0156 (0%)** | **19.5313 (+11%)** | **676 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|7.794 ms|0.0235 ms|226.5625|78.1250|3,765 KB|
| **New** |	|	| **6.989 ms (-10%)** | **0.0099 ms** | **226.5625 (0%)** | **78.1250 (0%)** | **3,764 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|11.860 ms|0.0355 ms|343.7500|171.8750|5,800 KB|
| **New** |	|	| **11.105 ms (-6%)** | **0.0396 ms** | **343.7500 (0%)** | **171.8750 (0%)** | **5,798 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|6.255 ms|0.0169 ms|187.5000|78.1250|3,090 KB|
| **New** |	|	| **5.788 ms (-7%)** | **0.0117 ms** | **187.5000 (0%)** | **78.1250 (0%)** | **3,091 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|1.109 ms|0.0021 ms|33.2031|13.6719|571 KB|
| **New** |	|	| **1.011 ms (-9%)** | **0.0028 ms** | **33.2031 (0%)** | **13.6719 (0%)** | **571 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|5.561 ms|0.0099 ms|171.8750|70.3125|2,857 KB|
| **New** |	|	| **5.133 ms (-8%)** | **0.0122 ms** | **171.8750 (0%)** | **70.3125 (0%)** | **2,856 KB (0%)** |