@adams85 2022-08-26
> new max for test code is 471

Above 400 we should be good.

Now we've reached the point when a preview release seems to be a good idea, haven't we? I mean the public API has solidified and the lib is kinda production ready. I expect bugfixes mainly from now on.