@lahma 2022-08-26
There's now new beta on NuGet: [3.0.0-beta-6](https://www.nuget.org/packages/Esprima/3.0.0-beta-6)